@@ -8,16 +8,19 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
+  cupertino_icons: ^1.0.3
+  dartz: ^0.10.0-nullsafety.2
   flutter:
     sdk: flutter
-  cupertino_icons: null
   flutter_rating_bar: ^4.0.0
   flutter_svg: ^0.22.0
-  get: ^4.1.4
+  flutter_typeahead: ^3.2.0
+  get: ^4.3.4
   google_fonts: ^2.1.0
+  http: ^0.13.3
 
 dev_dependencies:
   flutter_test:
