@@ -6,7 +6,7 @@ import 'package:get/get.dart';
 
 class TimeTablePage extends StatefulWidget {
   static const String routeName = '/timetablePage';
-  TimeTablePage({Key key}) : super(key: key);
+  TimeTablePage({Key? key}) : super(key: key);
 
   @override
   _TimeTablePageState createState() => _TimeTablePageState();
@@ -31,64 +31,61 @@ class _TimeTablePageState extends State<TimeTablePage> {
               icon: Icon(Icons.download))
         ],
       ),
-      body: SingleChildScrollView(
-        scrollDirection: Axis.vertical,
-        child: Column(
-          children: [
-            Container(
-              height: size.height * .78,
-              color: Colors.grey.shade100,
+      body: Column(
+        children: [
+          Container(
+            height: size.height * .77,
+            color: Colors.grey.shade100,
+            child: SingleChildScrollView(
               child: SingleChildScrollView(
-                child: SingleChildScrollView(
-                  scrollDirection: Axis.horizontal,
-                  child: Column(
-                    children: [
-                      TheoryHeader(),
-                      LabHeader(),
-                      WeekDayTT(
-                        day: 'Monday',
-                      ),
-                      WeekDayTT(
-                        day: 'Tuesday',
-                      ),
-                      WeekDayTT(
-                        day: 'Wednes\nday',
-                      ),
-                      WeekDayTT(
-                        day: 'Thrusday',
-                      ),
-                      WeekDayTT(
-                        day: 'Friday',
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-            ),
-            SizedBox(
-              height: 10,
-            ),
-            Padding(
-              padding: const EdgeInsets.all(8.0),
-              child: Container(
-                width: 300,
-                height: 50,
-                child: MaterialButton(
-                  color: KColor.primaryColor,
-                  onPressed: () => //First check login then review
-                      Get.toNamed(SignUpPage.routeName),
-                  child: Text(
-                    'Review Faculty',
-                    style: TextStyle(
-                      fontSize: 18,
-                      color: Colors.white,
+                scrollDirection: Axis.horizontal,
+                child: Column(
+                  children: [
+                    TheoryHeader(),
+                    LabHeader(),
+                    WeekDayTT(
+                      day: 'Monday',
                     ),
+                    WeekDayTT(
+                      day: 'Tuesday',
+                    ),
+                    WeekDayTT(
+                      day: 'Wednes\nday',
+                    ),
+                    WeekDayTT(
+                      day: 'Thrusday',
+                    ),
+                    WeekDayTT(
+                      day: 'Friday',
+                    ),
+                  ],
+                ),
+              ),
+            ),
+          ),
+          SizedBox(
+            height: 10,
+          ),
+          Padding(
+            padding: const EdgeInsets.all(8.0),
+            child: Container(
+              width: double.infinity,
+              height: size.height * .06,
+              child: MaterialButton(
+                color: KColor.primaryColor,
+                onPressed: () => //First check login then review
+                    Get.toNamed(SignUpPage.routeName),
+                child: Text(
+                  'Review Faculty',
+                  style: TextStyle(
+                    fontSize: 18,
+                    color: Colors.white,
                   ),
                 ),
               ),
-            )
-          ],
-        ),
+            ),
+          )
+        ],
       ),
     );
   }
